Rename hoc folder to hocs
